python_client/requirements.txt
grpcio==1.78.1
grpcio-tools==1.78.1
